refactor: reuse existing client response
Reuses the batch creation response in the Java client for process instance incident resolution.
Extends the existing response by comments and the batch operation type returned in the API response.
