--- clients/java/src/main/java/io/camunda/client/impl/response/BatchOperationCreatedResponseImpl.java
@@ -1,43 +1,0 @@
-/*
- * Copyright © 2017 camunda services GmbH (info@camunda.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.camunda.client.impl.response;
-
-import io.camunda.client.api.response.BatchOperationCreatedResponse;
-import io.camunda.client.api.search.enums.BatchOperationType;
-import io.camunda.client.impl.util.EnumUtil;
-import io.camunda.client.protocol.rest.BatchOperationCreatedResult;
-
-public final class BatchOperationCreatedResponseImpl implements BatchOperationCreatedResponse {
-
-  private final long batchOperationKey;
-  private final BatchOperationType batchOperationType;
-
-  public BatchOperationCreatedResponseImpl(final BatchOperationCreatedResult response) {
-    batchOperationKey = Long.parseLong(response.getBatchOperationKey());
-    batchOperationType =
-        EnumUtil.convert(response.getBatchOperationType(), BatchOperationType.class);
-  }
-
-  @Override
-  public long getBatchOperationKey() {
-    return batchOperationKey;
-  }
-
-  @Override
-  public BatchOperationType getBatchOperationType() {
-    return batchOperationType;
-  }
-}
