+++ clients/java/src/main/java/io/camunda/client/impl/response/CreateBatchOperationResponseImpl.java
@@ -16,19 +16,28 @@
 package io.camunda.client.impl.response;
 
 import io.camunda.client.api.response.CreateBatchOperationResponse;
+import io.camunda.client.api.search.enums.BatchOperationType;
+import io.camunda.client.impl.util.EnumUtil;
 import io.camunda.client.protocol.rest.BatchOperationCreatedResult;
 
 public class CreateBatchOperationResponseImpl implements CreateBatchOperationResponse {
 
-  private String batchOperationKey;
+  private final String batchOperationKey;
+  private final BatchOperationType batchOperationType;
+
+  public CreateBatchOperationResponseImpl(final BatchOperationCreatedResult response) {
+    batchOperationKey = response.getBatchOperationKey();
+    batchOperationType =
+        EnumUtil.convert(response.getBatchOperationType(), BatchOperationType.class);
+  }
 
   @Override
   public String getBatchOperationKey() {
     return batchOperationKey;
   }
 
-  public CreateBatchOperationResponseImpl setResponse(final BatchOperationCreatedResult response) {
-    batchOperationKey = response.getBatchOperationKey();
-    return this;
+  @Override
+  public BatchOperationType getBatchOperationType() {
+    return batchOperationType;
   }
 }
